feat: convert VIP into Regular and merged the vips
